requirements.txt
Flask==3.0.2
matplotlib==3.7.5
google.auth
requests

